// src/types/MoviesType.ts
export interface IMovieSearchData {
  imdbID: string;
  Title: string;
  Year: string;
  Poster: string;
}

export interface IMovieDetailsData {
  imdbID: string;
  Title: string;
  Year: string;
  Poster: string;
  Actors: string;
  Country: string;
  Genre: string;
  Plot: string;
  Released: string;
  Runtime: string;
  imdbRating: string;
}